Create OrdersModule
