--- app/actions/OrdersActionCreator.test.js
@@ -1,11 +1,0 @@
-// TODO: Read a bit more about how to test async actions. =)
-import 'react-native';
-// import React from 'react';
-
-describe('Orders actions', () => {
-	test('get orders', () => {
-	});
-
-	test('refresh orders', () => {
-	});
-});
