--- app/utils/constants/ActionTypes.js
@@ -1,7 +1,0 @@
-const ActionTypes = {
-	fetchOrders: 'FETCH_ORDERS',
-	ordersReceived: 'ORDERS_RECEIVED',
-	fetchingServerError: 'FETCHING_SERVER_ERROR'
-};
-
-export default ActionTypes;
